Create a Count Table with Percentages, Mean, and SD

This function takes a data frame, transforms it by computing counts, percentages, mean, and standard deviation for specified columns. It helps in summarizing the data to understand the distribution and variation.

Create a count table with percentages, mean, and standard deviation.

Usage

counting(df1)

counting(df1)

Arguments

df1

A data frame containing the data to be processed.

Value

A data frame with computed statistics, including counts, percentages, mean, and standard deviation.

A data frame with computed statistics.

Examples

df1 <- data.frame(A = c(-1, 2, 2, 3, -1), B = c(-1, 2, 3, -1, 2),
C = c(1, 2, -2, 3, -1), D = c(3, 2, 1, -1, -2), E = c(2, 3, 1, -1, -3))
result <- counting(df1)
print(result)
#> # A tibble: 4 × 16
#>   name   Mean    SD Total `Count -1` `% -1` `Count 2` `% 2` `Count 3` `% 3`
#>   <chr> <dbl> <dbl> <int>      <int>  <dbl>     <int> <dbl>     <int> <dbl>
#> 1 B       1    1.87     5          2     40         2    40         1    20
#> 2 C       0.6  2.07     5          1     20         1    20         1    20
#> 3 D       0.6  2.07     5          1     20         1    20         1    20
#> 4 E       0.4  2.41     5          1     20         1    20         1    20
#> # ℹ 6 more variables: `Count -2` <int>, `% -2` <dbl>, `Count 1` <int>,
#> #   `% 1` <dbl>, `Count -3` <int>, `% -3` <dbl>
df1 <- data.frame(A = c(-1, 2, 2, 3, -1), B = c(-1, 2, 3, -1, 2),
                  C = c(1, 2, -2, 3, -1), D = c(3, 2, 1, -1, -2), E = c(2, 3, 1, -1, -3))
counting(df1)
#> # A tibble: 4 × 16
#>   name   Mean    SD Total `Count -1` `% -1` `Count 2` `% 2` `Count 3` `% 3`
#>   <chr> <dbl> <dbl> <int>      <int>  <dbl>     <int> <dbl>     <int> <dbl>
#> 1 B       1    1.87     5          2     40         2    40         1    20
#> 2 C       0.6  2.07     5          1     20         1    20         1    20
#> 3 D       0.6  2.07     5          1     20         1    20         1    20
#> 4 E       0.4  2.41     5          1     20         1    20         1    20
#> # ℹ 6 more variables: `Count -2` <int>, `% -2` <dbl>, `Count 1` <int>,
#> #   `% 1` <dbl>, `Count -3` <int>, `% -3` <dbl>